refactoring and docstrings
