--- joeynmt/transformer.py
@@ -1,248 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import math
-import copy
-import numpy as np
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from torch import Tensor
-
-# This contains helper classes for the Transformer.
-# Source: http://nlp.seas.harvard.edu/2018/04/03/attention.html
-
-
-def clones(module, N):
-    """Produce N identical layers."""
-    return nn.ModuleList([copy.deepcopy(module) for _ in range(N)])
-
-
-# pylint: disable=arguments-differ,too-many-arguments
-class SublayerConnection(nn.Module):
-    """
-    Transformer residual connection
-    Applies sublayer to x, layer norm, dropout, then adds x.
-    """
-    def __init__(self, size, dropout):
-        super(SublayerConnection, self).__init__()
-        self.norm = nn.LayerNorm(size)
-        self.dropout = nn.Dropout(dropout)
-
-    def forward(self, x, sublayer):
-        return x + self.dropout(sublayer(self.norm(x)))
-
-
-def subsequent_mask(size):
-    """
-    Mask out subsequent positions.
-    """
-    attn_shape = (1, size, size)
-    subseq_mask = np.triu(np.ones(attn_shape), k=1).astype('uint8')
-    return torch.from_numpy(subseq_mask) == 0
-
-
-def attention(
-        query: Tensor,
-        key: Tensor,
-        value: Tensor,
-        mask: Tensor = None,
-        dropout: float = None):
-    """
-    Computes scaled dot-product attention
-
-    :param query:
-    :param key:
-    :param value:
-    :param mask:
-    :param dropout:
-    :return:
-    """
-    d_k = query.size(-1)
-    scores = torch.matmul(query, key.transpose(-2, -1)) \
-             / math.sqrt(d_k)
-    if mask is not None:
-        scores = scores.masked_fill(mask == 0, -1e9)
-    p_attn = F.softmax(scores, dim=-1)
-    if dropout is not None:
-        p_attn = dropout(p_attn)
-    return torch.matmul(p_attn, value), p_attn
-
-
-# pylint: disable=arguments-differ,too-many-arguments
-class MultiHeadedAttention(nn.Module):
-    def __init__(self,
-                 num_heads: int,
-                 hidden_size: int,
-                 dropout: float = 0.1):
-        """
-        Multi-headed attention
-
-        :param num_heads: number of attention heads
-        :param hidden_size: model size
-        :param dropout:
-        """
-        super(MultiHeadedAttention, self).__init__()
-        assert hidden_size % num_heads == 0, \
-            "hidden_size must be divisible by num_heads"
-
-        # We assume d_v always equals d_k
-        self.d_k = hidden_size // num_heads
-        self.h = num_heads
-        self.linears = clones(nn.Linear(hidden_size, hidden_size), 4)
-        self.attn = None
-        self.dropout = nn.Dropout(p=dropout)
-
-    def forward(self, query, key, value, mask=None):
-        """
-        Compute multi-headed attention.
-
-        :param query:
-        :param key:
-        :param value:
-        :param mask:
-        :return: new x
-        """
-
-        if mask is not None:
-            # Same mask applied to all h heads.
-            mask = mask.unsqueeze(1)
-        nbatches = query.size(0)
-
-        # 1) Do all the linear projections in batch from d_model => h x d_k
-        query, key, value = \
-            [l(x).view(nbatches, -1, self.h, self.d_k).transpose(1, 2)
-             for l, x in zip(self.linears, (query, key, value))]
-
-        # 2) Apply attention on all the projected vectors in batch.
-        x, self.attn = attention(query, key, value, mask=mask,
-                                 dropout=self.dropout)
-
-        # 3) "Concat" using a view and apply a final linear.
-        x = x.transpose(1, 2).contiguous() \
-            .view(nbatches, -1, self.h * self.d_k)
-        return self.linears[-1](x)
-
-
-class PositionwiseFeedForward(nn.Module):
-    """
-    Position-wise Feed-forward layer
-    Projects to d_ff and then back down to d_model.
-    """
-
-    def __init__(self, d_model, d_ff, dropout=0.1):
-        """
-        Initializes position-wise feed-forward layer.
-        :param d_model: dimensionality of the input.
-        :param d_ff: dimensionality of intermediate representation
-        :param dropout:
-        """
-        super(PositionwiseFeedForward, self).__init__()
-        self.layer = nn.Sequential(
-            nn.Linear(d_model, d_ff),
-            nn.ReLU(),
-            nn.Dropout(dropout),
-            nn.Linear(d_ff, d_model)
-        )
-
-    def forward(self, x):
-        return self.layer(x)
-
-
-class PositionalEncoding(nn.Module):
-    """
-    Pre-compute position encodings.
-    In forward pass, this adds the position-encodings to the
-    input for as many time steps as necessary.
-    """
-
-    def __init__(self, d_model, dropout, max_len=5000):
-        """
-
-        :param d_model: dimensionality of the model
-        :param dropout:
-        :param max_len: maximum length for which  to pre-compute PE
-        """
-        super(PositionalEncoding, self).__init__()
-        self.dropout = nn.Dropout(p=dropout)
-
-        # Compute the positional encodings once in log space.
-        pe = torch.zeros(max_len, d_model)
-        position = torch.arange(0, max_len, dtype=torch.float32).unsqueeze(1)
-        div_term = torch.exp(torch.arange(0, d_model, 2, dtype=torch.float32) *
-                             -(math.log(10000.0) / d_model))
-        pe[:, 0::2] = torch.sin(position * div_term)
-        pe[:, 1::2] = torch.cos(position * div_term)
-        pe = pe.unsqueeze(0)
-        self.register_buffer('pe', pe)
-
-    def forward(self, x):
-        """
-        Add PE to x.
-        :param x:
-        :return: x with position encodings added.
-        """
-        x = x + self.pe[:, :x.size(1)]
-        return self.dropout(x)
-
-
-class NoamScheduler:
-    """
-    The Noam learning rate scheduler used in "Attention is all you need"
-    See Eq. 3 in https://arxiv.org/pdf/1706.03762.pdf
-    """
-
-    def __init__(self, model_size, factor, warmup, optimizer):
-        self.optimizer = optimizer
-        self._step = 0
-        self.warmup = warmup
-        self.factor = factor
-        self.model_size = model_size
-        self._rate = 0
-
-    def step(self):
-        """Update parameters and rate"""
-        self._step += 1
-        rate = self.rate()
-        for p in self.optimizer.param_groups:
-            p['lr'] = rate
-        self._rate = rate
-
-    def rate(self, step=None):
-        """Implement `lrate` above"""
-        if step is None:
-            step = self._step
-        return self.factor * \
-               (self.model_size ** (-0.5) *
-                min(step ** (-0.5), step * self.warmup ** (-1.5)))
-
-    #pylint: disable=no-self-use
-    def state_dict(self):
-        return None
-
-
-class LabelSmoothing(nn.Module):
-    """
-    Label smoothing.
-    Used in Transformer.
-
-    """
-    def __init__(self, size, padding_idx, smoothing=0.0):
-        super(LabelSmoothing, self).__init__()
-        self.criterion = nn.KLDivLoss(reduction='sum')
-        self.padding_idx = padding_idx
-        self.confidence = 1.0 - smoothing
-        self.smoothing = smoothing
-        self.size = size
-        self.true_dist = None
-
-    def forward(self, x, target):
-        assert x.size(1) == self.size
-        true_dist = x.data.clone()
-        true_dist.fill_(self.smoothing / (self.size - 2))
-        true_dist.scatter_(1, target.data.unsqueeze(1), self.confidence)
-        true_dist[:, self.padding_idx] = 0
-        mask = torch.nonzero(target.data == self.padding_idx)
-        if mask.dim() > 0:
-            true_dist.index_fill_(0, mask.squeeze(), 0.0)
-        self.true_dist = true_dist
-        return self.criterion(x, true_dist)
